File reorganization

+++ Core/SNES/Debugger/NecDspDebugger.cpp
@@ -1,5 +1,4 @@
 #include "stdafx.h"
-#include "Emulator.h"
 #include "SNES/BaseCartridge.h"
 #include "SNES/MemoryManager.h"
 #include "SNES/Console.h"
@@ -13,7 +12,8 @@
 #include "Debugger/Debugger.h"
 #include "Debugger/MemoryAccessCounter.h"
 #include "Debugger/ExpressionEvaluator.h"
-#include "EmuSettings.h"
+#include "Shared/Emulator.h"
+#include "Shared/EmuSettings.h"
 #include "MemoryOperationType.h"
 
 NecDspDebugger::NecDspDebugger(Debugger* debugger)
@@ -60,7 +60,7 @@ void NecDspDebugger::ProcessRead(uint32_t addr, uint8_t value, MemoryOperationTy
 
 void NecDspDebugger::ProcessWrite(uint32_t addr, uint8_t value, MemoryOperationType type)
 {
-	AddressInfo addressInfo { addr, SnesMemoryType::DspDataRam }; //Writes never affect the DSP ROM
+	AddressInfo addressInfo { (int32_t)addr, SnesMemoryType::DspDataRam }; //Writes never affect the DSP ROM
 	MemoryOperationInfo operation { addr, value, type };
 	_debugger->ProcessBreakConditions(false, GetBreakpointManager(), operation, addressInfo);
 }
